improve structure and kafka events for everyone
