+++ app/src/main/java/io/apicurio/registry/events/kafka/KafkaEventSink.java
@@ -1,4 +1,19 @@
-package io.apicurio.registry.streams.events;
+/*
+ * Copyright 2020 Red Hat
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.apicurio.registry.events.kafka;
 
 import java.time.Instant;
 import java.util.Optional;
@@ -34,7 +49,7 @@ public class KafkaEventSink implements EventSink {
 
     @Inject
     @RegistryProperties(
-            value = {"registry.streams.common", "registry.streams.events"},
+            value = {"registry.events.kafka.config"},
             empties = {"ssl.endpoint.identification.algorithm="}
     )
     Properties producerProperties;
