+++ app/src/main/resources/application.properties
@@ -100,9 +100,6 @@ quarkus.index-dependency.jaxrs.artifact-id=jboss-jaxrs-api_2.1_spec
 #%prod.registry.streams.storage-producer.max.in.flight.requests.per.connection=5
 %prod.registry.streams.storage-producer.retries=3
 %prod.registry.streams.storage-producer.acks=all
-%prod.registry.streams.events.enable.idempotence=true
-%prod.registry.streams.events.retries=3
-%prod.registry.streams.events.acks=all
 
 # Asyncmem -- when used
 %prod.registry.asyncmem.delays.create=${ASYNC_DELAYS_CREATE:1000}
@@ -161,3 +158,13 @@ mp.openapi.servers=/api
 # compatibility with knative sink binding
 %dev.registry.events.ksink=${K_SINK:}
 %prod.registry.events.ksink=${K_SINK:}
+
+%dev.registry.events.kafka.config.bootstrap.servers=${bootstrap.servers:localhost:9092}
+%dev.registry.events.kafka.config.enable.idempotence=true
+%dev.registry.events.kafka.config.retries=3
+%dev.registry.events.kafka.config.acks=all
+
+%prod.registry.events.kafka.config.bootstrap.servers=${KAFKA_BOOTSTRAP_SERVERS}
+%prod.registry.events.kafka.config.enable.idempotence=true
+%prod.registry.events.kafka.config.retries=3
+%prod.registry.events.kafka.config.acks=all
