+++ app/src/test/resources/application.properties
@@ -49,3 +49,8 @@ h2.port=9123
 
 # UI Features
 %test.registry.ui.features.readOnly=${REGISTRY_UI_FEATURES_READONLY:false}
+
+%test.registry.events.kafka.config.bootstrap.servers=${bootstrap.servers:localhost:9092}
+%test.registry.events.kafka.config.enable.idempotence=true
+%test.registry.events.kafka.config.retries=3
+%test.registry.events.kafka.config.acks=all
